add tests

+++ src/main/kotlin/com/example/logic/admin/adminProducts/AdminProductsController.kt
@@ -3,9 +3,10 @@ package com.example.logic.admin.adminProducts
 import com.example.logic.authorization.database.ConstClient
 import com.example.logic.authorization.database.ConstProduct
 import com.example.logic.authorization.database.DatabaseHandler
+import com.example.logic.authorization.tables.Client
 import com.example.logic.authorization.tables.Product
-import com.example.logic.clientAccount.ClientAccountController
 import com.example.logic.logs.LogApp
+import com.sun.org.apache.xpath.internal.operations.Bool
 import tornadofx.Controller
 import java.lang.Exception
 import java.sql.SQLException
@@ -14,7 +15,9 @@ import java.util.logging.Level
 class AdminProductsController: Controller() {
     private val myLog = LogApp.getLogger()
 
-    fun changeNumOfProduct(productId: Int,newNumOfProduct: Int) {
+    fun changeNumOfProduct(productId: Int,newNumOfProduct: Int): Boolean {
+        if (!isProudctIdInList(productId)) return false
+
         val updateAmount = "UPDATE ${ConstProduct.PRODUCT_TABLE} SET ${ConstProduct.PRODUCT_NUMBER_IN_STOCK} " +
                 "= $newNumOfProduct " +
                 "WHERE ${ConstProduct.PRODUCT_ID} = $productId"
@@ -25,6 +28,7 @@ class AdminProductsController: Controller() {
 
             myLog.log(Level.INFO, "AdminProductsController: Кол-во продуктов с id=$productId успешно " +
                     "изменено на $newNumOfProduct")
+            if (preparedStatement != null) return true
         }catch (e: SQLException) {
             myLog.log(Level.SEVERE, "AdminProductsController: SQLException", e)
         }catch (e: ClassNotFoundException) {
@@ -32,9 +36,11 @@ class AdminProductsController: Controller() {
         }catch (e: Exception) {
             myLog.log(Level.SEVERE, "AdminProductsController: Exception", e)
         }
+        return false
     }
 
-    fun changeCostOfProduct(productId: Int, newCostOfProduct: Int) {
+    fun changeCostOfProduct(productId: Int, newCostOfProduct: Int): Boolean {
+        if (!isProudctIdInList(productId)) return false
         val updateAmount = "UPDATE ${ConstProduct.PRODUCT_TABLE} SET ${ConstProduct.PRODUCT_COST} " +
                 "= $newCostOfProduct " +
                 "WHERE ${ConstProduct.PRODUCT_ID} = $productId"
@@ -45,6 +51,7 @@ class AdminProductsController: Controller() {
 
             myLog.log(Level.INFO, "AdminProductsController: Стоимость продукта с id=$productId " +
                     "успешно изменена на $newCostOfProduct")
+            if (preparedStatement != null) return true
         }catch (e: SQLException) {
             myLog.log(Level.SEVERE, "AdminProductsController: SQLException", e)
         }catch (e: ClassNotFoundException) {
@@ -52,6 +59,7 @@ class AdminProductsController: Controller() {
         }catch (e: Exception) {
             myLog.log(Level.SEVERE, "AdminProductsController: Exception", e)
         }
+        return false
     }
 
     fun createNewProduct(product: Product) {
@@ -78,7 +86,8 @@ class AdminProductsController: Controller() {
         }
     }
 
-    fun deleteProduct(productId: Int) {
+    fun deleteProduct(productId: Int): Boolean {
+        if (!isProudctIdInList(productId)) return false
         val deleteProduct = "DELETE FROM ${ConstProduct.PRODUCT_TABLE} WHERE ${ConstProduct.PRODUCT_ID} = $productId"
 
         try {
@@ -87,6 +96,7 @@ class AdminProductsController: Controller() {
             preparedStatement.executeUpdate()
 
             myLog.log(Level.INFO, "AdminProductsController: Продукт с id=$productId успешно удален")
+            if (preparedStatement != null) return true
         }catch (e: SQLException) {
             myLog.log(Level.SEVERE, "AdminProductsController: SQLException", e)
         }catch (e: ClassNotFoundException) {
@@ -94,5 +104,45 @@ class AdminProductsController: Controller() {
         }catch (e: Exception) {
             myLog.log(Level.SEVERE, "AdminProductsController: Exception", e)
         }
+        return true
+    }
+
+    private fun getProductsId(): List<Int> {
+        val listOfClientId = mutableListOf<Int>()
+        val getProductId = "SELECT ${ConstProduct.PRODUCT_ID} FROM ${ConstProduct.PRODUCT_TABLE}"
+
+        try {
+            val preparedStatement = DatabaseHandler().getDbConnection()
+                    .prepareStatement(getProductId)
+            val resultSet = preparedStatement.executeQuery()
+
+            while (resultSet.next()) {
+                val columnCount = resultSet.metaData.columnCount
+                for (i in 1..columnCount) {
+                    val row = resultSet.getString(i).toInt()
+                    listOfClientId.add(row)
+                }
+            }
+
+            myLog.log(Level.INFO, "AdminProductsController: id продуктов успешно получены")
+        }catch (e: SQLException) {
+            myLog.log(Level.SEVERE, "AdminProductsController: SQLException", e)
+        }catch (e: ClassNotFoundException) {
+            myLog.log(Level.SEVERE, "AdminProductsController: ClassNotFoundException", e)
+        }catch (e: Exception) {
+            myLog.log(Level.SEVERE, "AdminProductsController: Exception", e)
+        }
+        return listOfClientId.toList()
+    }
+
+    private fun isProudctIdInList(productId: Int): Boolean {
+        var isProudctIdInList = false
+        getProductsId().forEach {
+            if (it == productId) {
+                isProudctIdInList = true
+                return@forEach
+            }
+        }
+        return isProudctIdInList
     }
 }
